requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
cfgv==3.3.1
click==8.1.3
coverage==6.4
distlib==0.3.4
Django==4.0.4
djangorestframework==3.13.1
filelock==3.7.0
flake8==4.0.1
identify==2.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
virtualenv==20.14.1
